add_sequence_tagging_api
