+++ examples/sequence_tagging/eval.py
@@ -25,14 +25,14 @@ import math
 import argparse
 import numpy as np
 
-from train import SeqTagging, ChunkEval, LacLoss
-from utils.configure import PDConfig
-from utils.check import check_gpu, check_version
-from reader import LacDataset, LacDataLoader
-
 work_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(os.path.join(work_dir, "../"))
+
 from hapi.model import set_device, Input
+from hapi.text.sequence_tagging import SeqTagging, ChunkEval, LacLoss
+from hapi.text.sequence_tagging import LacDataset, LacDataLoader
+from hapi.text.sequence_tagging import check_gpu, check_version
+from hapi.text.sequence_tagging import PDConfig
 
 import paddle.fluid as fluid
 from paddle.fluid.layers.utils import flatten
@@ -65,7 +65,10 @@ def main(args):
         device=place)
     model.load(args.init_from_checkpoint, skip_mismatch=True)
 
-    model.evaluate(eval_dataset.dataloader, batch_size=args.batch_size)
+    eval_result = model.evaluate(eval_dataset.dataloader, batch_size=args.batch_size)
+    print("precison: %.5f" % (eval_result["precision"][0]))
+    print("recall: %.5f" % (eval_result["recall"][0]))
+    print("F1: %.5f" % (eval_result["F1"][0]))
 
 
 if __name__ == '__main__':
