+++ hapi/text/sequence_tagging/__init__.py
@@ -1,0 +1,24 @@
+# Copyright (c) 2020 PaddlePaddle Authors. All Rights Reserved. 
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from hapi.text.sequence_tagging.reader import LacDataset as LacDataset
+from hapi.text.sequence_tagging.reader import LacDataLoader as LacDataLoader
+from hapi.text.sequence_tagging.sequence_tagging import SeqTagging as SeqTagging
+from hapi.text.sequence_tagging.sequence_tagging import Chunk_eval as Chunk_eval
+from hapi.text.sequence_tagging.sequence_tagging import LacLoss as LacLoss
+from hapi.text.sequence_tagging.sequence_tagging import ChunkEval as ChunkEval
+from hapi.text.sequence_tagging.utils.configure import PDConfig as PDConfig 
+from hapi.text.sequence_tagging.utils.check import check_gpu as check_gpu
+from hapi.text.sequence_tagging.utils.check import check_version as check_version
+
